LayoutTests/imported/w3c/web-platform-tests/dom/events/scrolling/scrollend-event-fired-after-snap-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented

Harness Error (FAIL), message = Unhandled rejection: unimplemented

FAIL Tests that scrollend is fired after scroll snap animation completion. promise_test: Unhandled rejection with value: object "Error: unimplemented"
FAIL Tests that scrollend is fired after fling snap animation completion. promise_test: Unhandled rejection with value: "Reaches the maximum frames."
